Remove JumpIfBranch instruction

+++ crates/rune/src/runtime/select.rs
@@ -11,18 +11,18 @@ use crate::runtime::{Future, Mut, Value, VmResult};
 /// A stored select.
 #[derive(Debug)]
 pub struct Select {
-    futures: FuturesUnordered<SelectFuture<i64, Mut<Future>>>,
+    futures: FuturesUnordered<SelectFuture<usize, Mut<Future>>>,
 }
 
 impl Select {
     /// Construct a new stored select.
-    pub(crate) fn new(futures: FuturesUnordered<SelectFuture<i64, Mut<Future>>>) -> Self {
+    pub(crate) fn new(futures: FuturesUnordered<SelectFuture<usize, Mut<Future>>>) -> Self {
         Self { futures }
     }
 }
 
 impl future::Future for Select {
-    type Output = VmResult<(i64, Value)>;
+    type Output = VmResult<(usize, Value)>;
 
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         let poll = Pin::new(&mut self.futures).poll_next(cx);
